SYS-2825 move existing bond extra tests
